Add progress bar for download and audio extract
Removes youtube-dl as a dependency.

A side effect of this commit is slightly slower song processing time.
ytdl-core is about 1s slower than youtube-dl but allows a progress bar
to be added. Moreover, ytdl-core does not require the youtube-dl binary
be bundled into the app.

Using the youtube-dl binary directly is twice as fast as this app, but
then there is no batch id3 tagging.